Move shared frontend code to new project
