+++ Rogue.sln
@@ -9,6 +9,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RogueCollection", "src\Rogu
 		{E85D0F19-D582-4240-8299-06ACCF51883B} = {E85D0F19-D582-4240-8299-06ACCF51883B}
 		{5512251B-12A0-4CFD-8A13-0E57974206F0} = {5512251B-12A0-4CFD-8A13-0E57974206F0}
 		{25CF5262-CC48-4884-A9DF-DC5534807CB4} = {25CF5262-CC48-4884-A9DF-DC5534807CB4}
+		{E0FD877C-9C21-44B8-A074-45998156A532} = {E0FD877C-9C21-44B8-A074-45998156A532}
 		{6130418E-34AB-444A-AAD5-D7269DB26009} = {6130418E-34AB-444A-AAD5-D7269DB26009}
 		{494B9490-564B-4EB9-A41D-B8A2A151115C} = {494B9490-564B-4EB9-A41D-B8A2A151115C}
 		{D82E17C1-009E-4EC6-A271-9A79DCAC6B42} = {D82E17C1-009E-4EC6-A271-9A79DCAC6B42}
@@ -61,6 +62,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Rogomatic_Player", "src\Rog
 	ProjectSection(ProjectDependencies) = postProject
 		{0F817ACA-2C76-41C3-BBFD-36D66A68E2E6} = {0F817ACA-2C76-41C3-BBFD-36D66A68E2E6}
 	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Frontend", "src\Shared\Frontend\Frontend.vcxproj", "{E0FD877C-9C21-44B8-A074-45998156A532}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -150,6 +153,14 @@ Global
 		{6130418E-34AB-444A-AAD5-D7269DB26009}.Release|x64.Build.0 = Release|x64
 		{6130418E-34AB-444A-AAD5-D7269DB26009}.Release|x86.ActiveCfg = Release|Win32
 		{6130418E-34AB-444A-AAD5-D7269DB26009}.Release|x86.Build.0 = Release|Win32
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Debug|x64.ActiveCfg = Debug|x64
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Debug|x64.Build.0 = Debug|x64
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Debug|x86.ActiveCfg = Debug|Win32
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Debug|x86.Build.0 = Debug|Win32
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Release|x64.ActiveCfg = Release|x64
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Release|x64.Build.0 = Release|x64
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Release|x86.ActiveCfg = Release|Win32
+		{E0FD877C-9C21-44B8-A074-45998156A532}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -165,5 +176,6 @@ Global
 		{5D94880B-C99D-887C-5219-9F7CBE21947C} = {864E2853-9C3F-482B-9677-50D4F5A0DDED}
 		{D82E17C1-009E-4EC6-A271-9A79DCAC6B42} = {91047729-F446-42B0-A7E2-A1FF1E8E621E}
 		{6130418E-34AB-444A-AAD5-D7269DB26009} = {0584C180-BB79-4D16-BD3F-ADE3A8FFC3A6}
+		{E0FD877C-9C21-44B8-A074-45998156A532} = {1EF0858A-7558-4B06-8441-5F6EB2E044FF}
 	EndGlobalSection
 EndGlobal
