+++ src/Shared/Frontend/environment.cpp
