+++ src/Shared/Frontend/game_config.cpp
