+++ src/Shared/Frontend/game_config.h
