+++ src/Shared/Frontend/replayable_input.h
@@ -4,7 +4,6 @@
 #include <vector>
 #include <condition_variable>
 #include <input_interface.h>
-#include <SDL.h>
 #include "game_config.h"
 
 struct Environment;
@@ -18,7 +17,6 @@ public:
     virtual char GetChar(bool block, bool for_string, bool *is_replay) override;
     virtual void Flush() override;
     virtual bool HasTypeahead() override;
-    virtual bool HandleEvent(const SDL_Event& e) = 0;
 
     void SaveGame(std::ostream& file);
     void RestoreGame(std::istream& file);
