+++ src/Shared/Frontend/utility.cpp
